adjusted styling, finalized edit/delete cost revamp via id instead of index
