+++ src/assets/fixedCostHeadings.ts
@@ -1,0 +1,10 @@
+const fixedCostHeadingArray = [
+    ['Name', 'text-center basis-6/24 pr-2 md:pr-6'],
+    ['Category', 'text-center basis-6/24 pr-2 md:pr-6'],
+    ['Amount ($)', 'text-center basis-3/24 pr-2 md:pr-6'],
+    ['Freq', 'text-center basis-6/24 pr-2 md:pr-6'],
+    ['Total', 'text-center basis-3/24 pr-2 md:pr-6'],
+    ['', 'text-center basis-3/24 pr-2 md:pr-6']
+  ]
+  
+  export default fixedCostHeadingArray
